feat : memberProductId

+++ src/components/Cart/CartPage.js
@@ -28,15 +28,15 @@ function CartPage() {
     fetchCartItems();
   }, [navigate]);
 
-  const handleDelete = async (productId) => {
+  const handleDelete = async (memberProductId) => {
     const authToken = localStorage.getItem('authToken');
     try {
-      const response = await axios.delete(`${process.env.REACT_APP_API_BASE_URL}/api/carts/products/${productId}`, {
+      const response = await axios.delete(`${process.env.REACT_APP_API_BASE_URL}/api/carts/products/${memberProductId}`, {
         headers: { Authorization: `Bearer ${authToken}` }
       });
 
       if (response.status === 204) { // 상태 코드 204가 반환되면 성공
-        setCartItems(cartItems.filter(item => item.id !== productId));
+        setCartItems(cartItems.filter(item => item.memberProductId !== memberProductId));
         alert('상품이 장바구니에서 삭제되었습니다.');
       } else {
         alert('상품 삭제에 실패했습니다.');
@@ -49,18 +49,28 @@ function CartPage() {
 
   const handleDeleteAll = async () => {
     const authToken = localStorage.getItem('authToken');
+    if (!authToken) {
+      alert('로그인이 필요합니다.');
+      navigate('/login');
+      return;
+    }
+
     try {
-      await axios.delete(`${process.env.REACT_APP_API_BASE_URL}/api/carts`, {
+      const response = await axios.delete(`${process.env.REACT_APP_API_BASE_URL}/api/carts`, {
         headers: { Authorization: `Bearer ${authToken}` }
       });
-      setCartItems([]);
-      alert('장바구니가 비워졌습니다.');
+
+      if (response.status === 204) { // 상태 코드 204가 반환되면 성공
+        setCartItems([]);
+        alert('장바구니가 비워졌습니다.');
+      } else {
+        alert('장바구니 비우기에 실패했습니다.');
+      }
     } catch (error) {
       console.error("전체 장바구니 삭제 중 오류 발생: ", error);
       alert('장바구니 전체 삭제 중 오류가 발생했습니다.');
     }
   };
-
   const handleGoToOrder = () => {
     navigate('/order');
   };
@@ -84,7 +94,7 @@ function CartPage() {
               <p>{item.title}</p>
               <p>수량: {item.productsQuantity}</p>
               <p>가격: {item.price}원</p>
-              <button onClick={() => handleDelete(item.id)}>삭제하기</button>
+              <button onClick={() => handleDelete(item.memberProductId)}>삭제하기</button>
             </div>
         ))}
         <button onClick={handleDeleteAll}>장바구니 비우기</button>
